refactor: major refactor, removed unneeded abstraction (#249)
The existing design was "too clever" and made the code hard to reason with. It previously created a wrapper around each TPLink device exposing all the characteristics it supported. It also created a HomeKit device "template" that would have characteristics that it wanted, and a process would match them up. The new design removes that extra layer of abstraction and makes the code much more clear. Also switched to the new Homebridge `onGet`/`onSet' methods that support Promises.